fix(tensor): #246 widen ASYM zeroPoint to int32 + [1,30] qBits ceiling
quantizeFloatToAsym stored round(min/scale) through an int16 narrowing
cast: for qBits >= 16 with min < 0 the true zeroPoint approaches
-(2^qBits - 1) (and overshoots it by min/(min - max) for all-negative
bands), so the cast silently wrapped and corrupted every
(code + zeroPoint) * scale dequantization.

- asymQConfig_t.zeroPoint widened int16_t -> int32_t; narrowing cast in
  deriveAsymGridFromMinMax removed (roundByMode already returns int32_t)
- ASYM qBits capped at [1, 30] in initAsymQConfig and re-checked in the
  deriveAsymGridFromMinMax funnel: at 31 the unsigned code ceiling
  (int32_t)(powf(2, qBits) - 1) rounds to 2^31 -> out-of-range cast in
  emitAsymChunk (unsigned twin of the #202 SYM_INT32 ceiling)
- deriveAsymGridFromMinMax fail-fasts when round(min/scale) would leave
  int32 (a narrow band far from zero does that at any qBits), keeping
  roundByMode float->int32 cast-safe
- ODTS wire + PPCA peek write/read zeroPoint as 4 bytes to match the
  widened field; the format-version break is owned by the #370 v2 bump
- tests: qBits=16 negative-band regression (exact zeroPoint -72817 +
  round-trip), two guard death tests; all zeroPoint carriers/asserts
  swept int16 -> int32

+++ src/serial/Serialize.c
@@ -28,8 +28,8 @@
  * one Record per layer (u8 tag + payload). Every count/dim/kernel field is u32
  * little-endian and every scalar goes through the checked SerialWire
  * primitives, so a model written on a 64-bit host loads bit-identically on
- * 32-bit MCU targets. ASYM zeroPoint is i32 LE on the wire (the in-memory
- * int16 widens separately via #246, keeping the format break to one bump). */
+ * 32-bit MCU targets. ASYM zeroPoint is i32 LE on the wire, matching the
+ * int32 in-memory field (#246). */
 #define SERIALIZE_MAGIC "ODTS"
 #define SERIALIZE_FORMAT_VERSION 2u
 
@@ -118,9 +118,7 @@ static void serializeQConfig(quantization_t *q, FILE *f) {
         serialWriteF32LE(asymQC->scale, f);
         serialWriteU8(asymQC->qBits, f);
         serialWriteU8((uint8_t)asymQC->roundingMode, f);
-        /* i32 on the wire although the field is (still) int16_t — #246 widens
-         * the in-memory side losslessly without another format break. */
-        serialWriteI32LE((int32_t)asymQC->zeroPoint, f);
+        serialWriteI32LE(asymQC->zeroPoint, f);
         break;
     }
     default:
